Extract OwnerUpgradesView
